requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
diffusers==0.23.1
transformers==4.35.2
torch>=2.2.0
accelerate==0.24.1
pillow==10.1.0
huggingface-hub==0.16.4 